+++ src/main/java/Main.java
@@ -8,20 +8,20 @@ public class Main {
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
         System.out.println("Enter number of towers:");
-        int numberOfTowers = scanner.nextInt();
+        byte numberOfTowers = scanner.nextByte();
         System.out.println("Enter number of disks:");
         int numberOfDisks = scanner.nextInt();
 
         //create initial scenarios
         List<Tower> towers = new ArrayList<>();
-        for (int i = 0; i < numberOfTowers; i++) {
+        for (byte i = 0; i < numberOfTowers; i++) {
             towers.add(new Tower(i));
         }
 
         //place all disks on first tower
         for (int i = numberOfDisks - 1; i >= 0; i--) {
             Tower sourceTower = towers.get(0);
-            sourceTower.putDisk(new Disk(i));
+            sourceTower.putDisk(new Disk((byte)i));
         }
 
         List<Scenario> scenarioList = new ArrayList<>();
